[Fix] compoent refactoring in HoverDetail
